Fixed naming convention to pass smoke test
