+++ bci_essentials/classification/__init__.py
@@ -5,10 +5,10 @@
 """
 
 from .generic_classifier import Generic_classifier
-from .erp_rg_classifier import ERP_rg_classifier
-from .mi_classifier import MI_classifier
-from .ssvep_riemmanian_mdm_classifier import SSVEP_riemannian_mdm_classifier
-from .ssvep_basic_tf_classifier import SSVEP_basic_tf_classifier
+from .erp_rg_classifier import ERP_rg_classifier as erp_rg_classifier
+from .mi_classifier import MI_classifier as mi_classifier
+from .ssvep_riemannian_mdm_classifier import SSVEP_riemannian_mdm_classifier as ssvep_riemannian_mdm_classifier
+from .ssvep_basic_tf_classifier import SSVEP_basic_tf_classifier as ssvep_basic_tf_classifier
 from .null_classifier import Null_classifier
 from .switch_mdm_classifier import Switch_mdm_classifier
 from .switch_deep_classifier import Switch_deep_classifier
