+++ bci_essentials/classification/ssvep_basic_tf_classifier.py
@@ -6,7 +6,7 @@ from scipy import signal
 
 # Custom libraries
 sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)),os.pardir))
-from classifiers.generic_classifier import Generic_classifier
+from classification.generic_classifier import Generic_classifier
 from bci_essentials.visuals import *
 from bci_essentials.signal_processing import *
 from bci_essentials.channel_selection import *
