+++ bci_essentials/classification/switch_deep_classifier.py
@@ -9,7 +9,7 @@ import tensorflow as tf
 
 # Custom libraries
 sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)),os.pardir))
-from classifiers.generic_classifier import Generic_classifier
+from classification.generic_classifier import Generic_classifier
 from bci_essentials.visuals import *
 from bci_essentials.signal_processing import *
 from bci_essentials.channel_selection import *
