+++ examples/ssvep_offline_test_tf.py
@@ -11,6 +11,7 @@ sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)),os.pardi
 
 # # from src.bci_data import *
 from bci_essentials.bci_data import *
+from bci_essentials.classification import ssvep_basic_tf_classifier
 
 # import
 import matplotlib.pyplot as plt
@@ -20,7 +21,7 @@ import matplotlib.pyplot as plt
 test_ssvep = EEG_data()
 
 # Define the classifier
-test_ssvep.classifier = ssvep_basic_classifier_tf(subset=[])
+test_ssvep.classifier = ssvep_basic_tf_classifier(subset=[])
 
 # Load from xdf into erp_data format
 test_ssvep.load_offline_eeg_data(filename = "examples\data\ssvep_example.xdf", format='xdf')
