+++ tests/test_smoke.py
@@ -1,5 +1,8 @@
 import unittest
+import os
+import sys
 
+sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)),os.pardir))
 from bci_essentials.bci_data import EEG_data, ERP_data
 from bci_essentials.classification import (
     mi_classifier,
